Periodic update.

+++ Personal/dev/opt/libcx/include/libcx/conf/prelude.hh
@@ -12,13 +12,13 @@ template<bool B> predicate kpred = B;
 template<isize bit>
 onedef cons isize BIT = 1ull << bit;
 
-using ErrorCode = u8;
-onedef cons u8 Error_None = 0;
-onedef cons u8 Operation_Fail = 1;
-onedef cons u8 Invalid_Ptr = 2;
-onedef cons u8 Invalid_Arg = 3;
-onedef cons u8 Bad_Alloc = 4;
-onedef cons u8 Invalid_Mode = 5;
+using ErrorCode = i32;
+onedef cons i32 Error_None = 0;
+onedef cons i32 Operation_Fail = 1;
+onedef cons i32 Invalid_Ptr = 2;
+onedef cons i32 Invalid_Arg = 3;
+onedef cons i32 Bad_Alloc = 4;
+onedef cons i32 Invalid_Mode = 5;
 
 }
